Move parser to its own package
