+++ static_analyzer.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 
 	"github.com/google/go-jsonnet/ast"
+	"github.com/google/go-jsonnet/parser"
 )
 
 type analysisState struct {
@@ -77,7 +78,7 @@ func analyzeVisit(a ast.Node, inObject bool, vars ast.IdentifierSet) error {
 		//nothing to do here
 	case *ast.SuperIndex:
 		if !inObject {
-			return MakeStaticError("Can't use super outside of an object.", *a.Loc())
+			return parser.MakeStaticError("Can't use super outside of an object.", *a.Loc())
 		}
 		visitNext(a.Index, inObject, vars, s)
 	case *ast.Index:
@@ -121,13 +122,13 @@ func analyzeVisit(a ast.Node, inObject bool, vars ast.IdentifierSet) error {
 		panic("Comprehensions not supported yet")
 	case *ast.Self:
 		if !inObject {
-			return MakeStaticError("Can't use self outside of an object.", *a.Loc())
+			return parser.MakeStaticError("Can't use self outside of an object.", *a.Loc())
 		}
 	case *ast.Unary:
 		visitNext(a.Expr, inObject, vars, s)
 	case *ast.Var:
 		if !vars.Contains(a.Id) {
-			return MakeStaticError(fmt.Sprintf("Unknown variable: %v", a.Id), *a.Loc())
+			return parser.MakeStaticError(fmt.Sprintf("Unknown variable: %v", a.Id), *a.Loc())
 		}
 		s.freeVars.Add(a.Id)
 	default:
